Custom local mods: CDJ jog feel, locked Instant-Pad FX unit, 4-deck layering, Color FX
Disclaimer (AI-generated): this commit was authored autonomously by an AI coding
agent at the repository owner's request; the human owner reviewed the diff before
publishing. Not intended for upstream contribution.

- controllers/scripting/legacy: global scratch-tick scaling x2.5 with >10-tick
  fling boost x1.3 (accumulator widened to double; SCOPE WARNING documented)
- engine/controls/ratecontrol: pitch-bend jog sensitivity member, startup read of
  [Controls] JogSensitivity (default 2.5), fling boost on raw value
- effects: freeze EffectUnit4 continuous surfaces (parameter/button_parameter,
  meta, super1, mix pinned at persisted value); enables, routing and effect
  loading remain operational for pad triggering
- res/controllers Numark Mixtrack (Pro) 3: WHEEL-hold deck layer swap 1<->3 /
  2<->4, force num_decks=4, layer-safe PADLoop/PFL/slip resolution, ColorFX
  reverb blend on filter knob above center

Disclaimer (AI-generated): end of autonomously authored commit text.

+++ src/controllers/scripting/legacy/controllerscriptinterfacelegacy.h
@@ -140,7 +140,8 @@ class ControllerScriptInterfaceLegacy : public QObject {
     };
     QHash<int, TimerInfo> m_timers;
 
-    QVarLengthArray<int> m_intervalAccumulator;
+    // === CUSTOM MOD (dynamic-jog): int -> double so scaled CDJ ticks accumulate untruncated ===
+    QVarLengthArray<double> m_intervalAccumulator;
     QVarLengthArray<mixxx::Duration> m_lastMovement;
     QVarLengthArray<double> m_dx, m_rampTo, m_rampFactor;
     QVarLengthArray<bool> m_ramp, m_brakeActive, m_spinbackActive, m_softStartActive;
